Allow for non-i32 arrays

+++ runner.cpp
@@ -2,7 +2,7 @@
 
 extern "C" {
     void STELLAR_ENTRY();
-    int getValue(int);
+    float getValue(int);
 }
 
 int main() {
